app/stockmonitor/serializers.py
from rest_framework import serializers
from core.models import WatchList


class WatchListSerializer(serializers.ModelSerializer):
    class Meta:
        model = WatchList
        fields = '__all__'
