requirements.txt
Django>=3.0.6,<3.1.0
djangorestframework==3.11.0
flake8>=3.8.2<3.9.0
psycopg2>=2.8.5<2.9.0
unittest-xml-reporting>=3.0.2,<3.1.0
django-rest-knox==4.1.0